Implement log in and inject identity
(1) Log in: If the username and password are correct,
return an access token.

(2) Identity injection: All routes get a req.identity
object, which will provide information on the user with
the specified access token (in query or body). If no
token is specified, req.identity will set the login type
to "guest" and requesting the username or uid will throw
an error.