Use only Authorization header and body to pass token
Passing token via query string is deemed unsafe, so we
get rid of that. Now the token will only be read from
Authorization header and request body. Header is preferred.

The request helper also automatically converts a `token`
param into the Authorization header, so all the request()
fetches are unaffected and will work the same as before.